put some testing framework in place
