+++ runner.go
@@ -58,7 +58,7 @@ type statusWithError struct {
 	err    error
 }
 
-func updateRunStatus(c *cli.Context, runID int, t *time.Ticker, resChan chan statusWithError) {
+func updateRunStatus(c cliContext, runID int, t *time.Ticker, resChan chan statusWithError) {
 	for {
 		// Wait for tick
 		<-t.C
@@ -81,12 +81,6 @@ func updateRunStatus(c *cli.Context, runID int, t *time.Ticker, resChan chan sta
 			resChan <- statusWithError{status: newStatus, err: nil}
 		}
 	}
-}
-
-type cliContext interface {
-	String(flag string) (val string)
-	StringSlice(flag string) (vals []string)
-	Args() (args cli.Args)
 }
 
 // makeRunParams parses and validates command line arguments + options
